better types

+++ include/stddef.h
@@ -2,12 +2,13 @@
 #define STDDEF_H
 
 #include <sys/types.h>
+#include <sys/type.h>
 
 #ifndef NULL
 #define NULL (void *)0
 #endif
 
-typedef signed long ptrdiff_t;
+typedef signed long long ptrdiff_t;
 
 typedef char wchar_t;
